% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{check_multicore}
\alias{check_multicore}
\title{Check Multicore}
\usage{
check_multicore(ncores)
}
\arguments{
\item{ncores}{number of cores}
}
\value{
the number of cores
}
\description{
Makes sure that multicore is not activated 
when using windows based OS
}
\keyword{internal}
