% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{clean_etr}
\alias{clean_etr}
\title{Clean ETR pixels}
\usage{
clean_etr(fv_fm_file, light_value, abso)
}
\arguments{
\item{fv_fm_file}{name of fv_fm file to be cleaned}

\item{light_value}{one light step value}

\item{abso}{absorbance value}
}
\value{
cleaned FvFm file
}
\description{
Some of the Fv/Fm will be noise pixels.
The second line of each ETR object replaces these pixels with 
NA values to be skipped during further calculations.
}
\keyword{internal}
