% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{clean_fvfm_file}
\alias{clean_fvfm_file}
\title{Clean Fv,Fm file}
\usage{
clean_fvfm_file(fvfm_file)
}
\arguments{
\item{fvfm_file}{a character vector with the full path
name of the fv-fm file}
}
\value{
cleaned fv-fm file
}
\description{
removes Fv_Fm values that are either negative or
}
\keyword{internal}
