% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_rlc}
\alias{create_rlc}
\title{Creates Rapid Light Curve}
\usage{
create_rlc(x2, y2, ncol)
}
\arguments{
\item{x2}{number of ...}

\item{y2}{number of ...}

\item{ncol}{number of columns of big matrix}
}
\description{
Creates a bigmemory object to store all the results
}
\keyword{internal}
