% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{etr_bigmatrix}
\alias{etr_bigmatrix}
\title{ETR bigmatrix object}
\usage{
etr_bigmatrix(path, total, light)
}
\arguments{
\item{path}{input directory}

\item{total}{number of rows}

\item{light}{steps}
}
\description{
creates an empty big.matrix object to store 
the rETR values of all the images it uses the bigmemory library
}
\keyword{internal}
