% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{etr_convergence}
\alias{etr_convergence}
\title{ETR Convergence}
\usage{
etr_convergence(etr_sim, rlc_params, pe3, model = "jassby")
}
\arguments{
\item{etr_sim}{etr}

\item{rlc_params}{RLC parameters}

\item{pe3}{data frame with light and etr}

\item{model}{name of fitted model}
}
\description{
update RLC parameters when etr converges to zero
}
\keyword{internal}
