% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{export_images}
\alias{export_images}
\title{Export Images}
\usage{
export_images(rlc_matrix, x2, model = "jassby")
}
\arguments{
\item{rlc_matrix}{RLC big matrix object}

\item{x2}{number of columns}

\item{model}{name of fitted model}
}
\description{
Export images as text files that can be imported by 
ImageJ as text Images
}
\keyword{internal}
