% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{get_absorbance}
\alias{get_absorbance}
\title{Calculate Absorbance}
\usage{
get_absorbance(path, ab)
}
\arguments{
\item{path}{a character vector with the full path name of the
directory where file fv-fm files are located}

\item{ab}{logical indicating whether absorbance is present}
}
\description{
Estimates rETR using absorbance information when
it is available
}
\keyword{internal}
