% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_mnb}
\alias{get_mnb}
\title{Get Mean Normalized Bias}
\usage{
get_mnb(predicted, pe3)
}
\arguments{
\item{predicted}{predicted value from \code{\link{get_prediction}}}

\item{pe3}{data frame with light and etr}
}
\description{
Mean normalized bias in percentage
}
\keyword{internal}
