% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_pixels}
\alias{get_pixels}
\title{Get Pixels}
\usage{
get_pixels(rlc_matrix, x2, y2)
}
\arguments{
\item{rlc_matrix}{RLC big matrix object}

\item{x2}{number of rows}

\item{y2}{number fo columns}
}
\description{
Calculates different types of pixels
}
\keyword{internal}
