% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_prediction}
\alias{get_prediction}
\title{Get Prediction}
\usage{
get_prediction(rlc_params, pe3, model = "jassby")
}
\arguments{
\item{pe3}{data frame with light and etr}

\item{model}{name of the fitted model}

\item{params}{list of RLC parameters}
}
\value{
predicted value
}
\description{
Compute predicted value
}
\keyword{internal}
