% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{get_retr}
\alias{get_retr}
\title{Calculate rETR values}
\usage{
get_retr(fv_fm, light, abso)
}
\arguments{
\item{fv_fm}{list with cleaned fv_fm files}

\item{light}{numeric vector of light steps}

\item{abso}{absorbance value}
}
\value{
list with etr matrices
}
\description{
Cycle to calculate the rETR values

Some of the Fv/Fm will be noise pixels.
The second line of each ETR object replaces these pixels with 
NA values to be skipped during further calculations.
}
\keyword{internal}
