% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_rmse}
\alias{get_rmse}
\title{Get Residual mean square error (RMSE)}
\usage{
get_rmse(predicted, pe3)
}
\arguments{
\item{predicted}{predicted value from \code{\link{get_prediction}}}

\item{pe3}{data frame with light and etr}
}
\description{
Residual mean square error (RMSE)
}
\keyword{internal}
