% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_wssr}
\alias{get_wssr}
\title{Weighted sum-of-squared-residuals}
\usage{
get_wssr(predicted, pe3)
}
\arguments{
\item{predicted}{predicted value from \code{\link{get_prediction}}}

\item{pe3}{data frame with light and etr}
}
\description{
Weighted sum-of-squared-residuals
}
\keyword{internal}
