% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{image_dimension}
\alias{image_dimension}
\title{Image Dimension}
\usage{
image_dimension(etr_matrix)
}
\arguments{
\item{etr_matrix}{image matrix}
}
\value{
list with \code{x2}, \code{y2}, and \code{total} values
}
\description{
Get image dimensions as total number of pixels
}
\keyword{internal}
