% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{import_fvfm_files}
\alias{import_fvfm_files}
\title{Import Fv,Fm files}
\usage{
import_fvfm_files(path, light)
}
\arguments{
\item{path}{a character vector with the full path name of the
directory where fv-fm files are located}

\item{light}{numeric vector of light steps}
}
\description{
Imports the Fv-Fm files
}
\keyword{internal}
