% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{import_light_steps}
\alias{import_light_steps}
\title{Import Light Steps File}
\usage{
import_light_steps(path, sep = ",")
}
\arguments{
\item{path}{a character vector with the full path name of the
directory where file light.txt is located}

\item{sep}{value separator character; comma by default}
}
\value{
numeric vector with light steps
}
\description{
Imports the file with the light curve light steps
}
\details{
The file to be imported needs to be named "light.txt"
By default, light values must be separated by a comma
}
\keyword{internal}
