% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{initialize_rlc_params}
\alias{initialize_rlc_params}
\title{Initialize RLC parameters}
\usage{
initialize_rlc_params(model = "jassby")
}
\arguments{
\item{model}{name of the fitted model}
}
\value{
list with various parameters
}
\description{
Creates default values to be updated during the fitting
process
}
\keyword{internal}
