% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mean_normalized_bias}
\alias{mean_normalized_bias}
\title{Mean Normalized Bias}
\usage{
mean_normalized_bias(predicted, pe3)
}
\arguments{
\item{predicted}{predicted value from \code{\link{get_prediction}}}

\item{pe3}{data frame with light and etr}
}
\description{
Mean normalized bias in percentage
}
\keyword{internal}
