% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{platt_model}
\alias{platt_model}
\title{Platt Model}
\usage{
platt_model(imgdim, etr_total, ncores = 1, plots = FALSE, alpha = 0.4,
  beta = 15, Ps = 15000, output_path = ".")
}
\arguments{
\item{imgdim}{list with image dimensions and light values
(output from \code{\link{import_data}})}

\item{etr_total}{etr big matrix object}

\item{ncores}{number of cores for parallel computing}

\item{plots}{should plots be printed during the fit? Only works when \code{ncores = 1}}

\item{alpha}{0.35 by default}

\item{beta}{0 by default}

\item{Ps}{40 by default}

\item{output_path}{character vector indicating full path where output files
will be saved}
}
\description{
Fits the light-photosynthesis curve model suggested by Platt, T., Gallegos, C.L., Harrison, W.G., 1980. Photoinhibition of photosynthesis in natural assemblages of marine phytoplankton. J. Mar. Res. 38, 687-701.
}
\details{
multicore will crash if any attemps are made to plot
anything during processing. Setting \code{plots} to \code{TRUE} when
\code{ncores > 1} will only plot the maps at the end of model fitting.
It is advised that a first run is carried out with a subset of
the image to check model fitting plotting the fit curve and
using only one core.
It creates a bigmemory object to store all the results
}
