% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print_completeness}
\alias{print_completeness}
\title{Print completeness}
\usage{
print_completeness(ncores, i, total)
}
\arguments{
\item{ncores}{number of cores}

\item{i}{counter}

\item{total}{number of pixels}
}
\description{
Print completeness percentage in the R console if using
just one core when multiple cores are used it is not possible to plot the
percentage because of the multiple R instances running,
thus a message is printed
}
\keyword{internal}
