% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{print_pixel_messages}
\alias{print_pixel_messages}
\title{Print Pixel Messages}
\usage{
print_pixel_messages(pixels_info)
}
\arguments{
\item{pixels_info}{information}
}
\description{
Prints messages about pixels
}
\keyword{internal}
