% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{res_mean_square_error}
\alias{res_mean_square_error}
\title{Residual Mean Square Error}
\usage{
res_mean_square_error(predicted, pe3)
}
\arguments{
\item{predicted}{predicted value from \code{\link{get_prediction}}}

\item{pe3}{data frame with light and etr}
}
\description{
Residual Mean Square Error
}
\keyword{internal}
