% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_1.R
\name{shiny1}
\alias{shiny1}
\title{shiny app to input a file and read in light levels}
\usage{
shiny1()
}
\value{
Returns an object (Raster brick?) with all the images from the imaging PAM file (ADD DETAILS)
}
\description{
shiny app to input a file and read in light levels (COMPLETE)
}
\examples{


}
\keyword{external}
