% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny2.R
\name{shiny2}
\alias{shiny2}
\title{shiny app to select the region of interest to be analysed}
\usage{
shiny2(my.data)
}
\value{
Returns an object (TYPE?) with all the PSII images and the rETR images cropped to the region of interest selected by the user
}
\description{
shiny app to select the region of interest to be analysed, it also calculates PSII efficiency images and the equivalent rETR images
}
\examples{


}
\keyword{external}
