% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny3.R
\name{shiny3}
\alias{shiny3}
\title{shiny app to select the irradiance-ETR curve to fit and calculate the parameter images}
\usage{
shiny3(my.data)
}
\value{
Returns an object (TYPE?) with all the rlc parameter images (e.g.alpha','etrmax', 'eopt', 'Ek')
}
\description{
shiny app to select the irradiance-ETR curve to fit and calculate the parameter images
}
\examples{


}
\keyword{external}
