% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny4.R
\name{shiny4}
\alias{shiny4}
\title{shiny app to view the results}
\usage{
shiny4(my.data)
}
\value{
It does not return any object, it for visualization only
}
\description{
shiny app to view the results
}
\examples{


}
\keyword{external}
