% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-files.R
\name{transform_etr}
\alias{transform_etr}
\title{Transform ETR}
\usage{
transform_etr(etr_total, list_etr, img_dim, path)
}
\arguments{
\item{etr_total}{big matrix object}

\item{list_etr}{list with etr matrices}

\item{img_dim}{image dimensions}
}
\description{
Cycle to fill the big.matrix object with
the rETR values with big images.
}
\details{
This will take same time to execute but having
all the rETR values in one big.matrix object facilitates the
calculations using \code{\link{mclapply}}, i.e. paralell computing.
}
\keyword{internal}
