% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_master.R
\name{pamfit}
\alias{pamfit}
\title{shiny app calculate images of RLC parameters}
\usage{
pamfit()
}
\value{
Returns a list with the images of the RLC parameters
}
\description{
shiny app calculate images of RLC parameters (add bigger description)
}
\examples{


}
\keyword{external}
